pyproject.toml
[project]
name = "updated-langchain-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
]
